Move TLS module up a level.
